+++ test/protocol/rack/adapter/rack3.rb
@@ -57,7 +57,7 @@ describe Protocol::Rack::Adapter::Rack3 do
 				let(:app) {->(env){[200, {}, fake_file]}}
 				
 				it "should not modify partial responses" do
-					expect(response.body).to be(:kind_of?, Protocol::Rack::Body::Enumerable)
+					expect(response.body).to be(:kind_of?, Protocol::HTTP::Body::Buffered)
 				end
 			end
 		end